Fixes to remove compiler warnings
Implement fixes for all warnings reported by compiler (GCC 4.3.4 with
-W -Wall flags). These fall into three main categories

- Local scope unused variables, fix by deletion or ignoring return
parameters.

- Unused function arguments, fix by commenting out argument name. This
helps to retain the context of the argument.

- Wrong type comparison (e.g. int -> size_t), fix by using correct type
or cast where behaviour uncertain.

+++ source/bxgenvtx/include/genvtx/polycone_vg.h
@@ -70,7 +70,7 @@ namespace genvtx {
 
     void set_surface (int surface_mask_);
 
-    void set_active_frustrum(int index_, bool active_);
+    void set_active_frustrum(size_t index_, bool active_);
 
     void set_polycone (const geomtools::polycone & polycone_);
